+++ src/app/api/specs/delete/route.js
@@ -1,17 +1,22 @@
+import { auth } from "@clerk/nextjs";
 import { NextResponse } from "next/server";
 import prisma from "@/lib/db";
 
 export async function POST(request) {
-  console.log(request);
-  try {
-    const { id } = await request.json();
+  const { userId } = auth();
+  if (userId)
+    try {
+      const { id } = await request.json();
 
-    await prisma.specification.delete({
-      where: { id: id },
-    });
+      await prisma.specification.delete({
+        where: { id: id },
+      });
 
-    return NextResponse.json({ message: "specification deleted successfully" });
-  } catch (err) {
-    return NextResponse.json({ message: "fail" });
-  }
+      return NextResponse.json({
+        message: "specification deleted successfully",
+      });
+    } catch (err) {
+      return NextResponse.json({ message: "fail" });
+    }
+  return new NextResponse("Unauthorized", { status: 401 });
 }
